+++ CHANGELOG.md
@@ -1,8 +1,15 @@
-# Unreleased
+###unreleased
+
+### Bug Fixes
+* **terra-node-tree** Fixed a bug concerning the deletion of a node on the first level
+* **TerraBreadcrumbsService** support to handle queryParams in url 
+
+<a name="3.0.0-beta.16"></a>
+# 3.0.0-beta.16 (29.11.2018)
 
 ### Breaking changes
 * **terra-tag** `inputBadge` is now deprecated. Please use `name` instead.
-* **TerraBreadcrumbsService** support to handle queryParams in url 
+* **terra-data-table** removed `inputIsSortable` due to redundancy.
 
 ### Bug Fixes
 * **terra-data-table** 
@@ -14,6 +21,7 @@
     - New input `inputIsSmall`
     - Added the ability to add a divider button to the 'inputOptions'
     - Fixed corresponding CSS
+* **terra-data-table.service** added `resetSortParams()` to be able to reset sorting params.
     
 <a name="3.0.0-beta.15"></a>
 # 3.0.0-beta.15 (21.11.2018)